requirements.txt
Babel==1.3
MySQL-python==1.2.4
Paste==1.7.5.1
PasteDeploy==1.5.0
PasteScript==1.7.5
#SQLAlchemy==0.8.3
argparse==1.2.1
iso8601==0.1.8
netaddr==0.7.10
oslo.config==1.2.1
pbr==0.5.23
python-keystoneclient==0.4.1
requests==2.0.1
six==1.4.1
wsgiref==0.1.2
